Переработан интерфейс программы.
