--- src/Main.java
@@ -1,9 +1,0 @@
-import objects.ApplicationRunner;
-import objects.Mode;
-
-public class Main {
-
-    public static void main(String[] args) {
-        ApplicationRunner.run(Mode.OBJECT);
-    }
-}
