+++ src/menu/Command.java
@@ -1,0 +1,7 @@
+package menu;
+
+public interface Command {
+    
+    Command execute();
+    
+}
