+++ src/menu/commands/tasks/calculator/Executor.java
@@ -1,0 +1,5 @@
+package menu.commands.tasks.calculator;
+
+public enum Executor {
+    IVAN
+}
