+++ src/menu/commands/tasks/calculator/Workplace.java
@@ -1,0 +1,6 @@
+package menu.commands.tasks.calculator;
+
+public interface Workplace {
+
+    void start();
+}
