+++ src/menu/commands/tasks/calculator/ivan/Calculator.java
@@ -1,0 +1,8 @@
+package menu.commands.tasks.calculator.ivan;
+
+class Calculator {
+
+    int sum(int firstOperand, int secondOperand) {
+        return  firstOperand + secondOperand;
+    }
+}
