+++ src/menu/commands/tasks/objects/exercises/Exercise.java
@@ -1,0 +1,7 @@
+package menu.commands.tasks.objects.exercises;
+
+public enum Exercise {
+    DNS,
+    FACTORY,
+    SCHOOL
+}
