+++ src/menu/commands/tasks/objects/exercises/Organization.java
@@ -1,0 +1,5 @@
+package menu.commands.tasks.objects.exercises;
+
+public abstract class Organization {
+    public abstract void start();
+}
