+++ src/menu/commands/tasks/objects/readme.md
@@ -1,0 +1,19 @@
+Необходимо составить объектную модель любой сложной структуры (Рок группы, завода, университета и т.п.) 
+создав не меньше трех взаимосязанных объектов, имеющих друг с другом общий суперкласс.
+
+Структура задания:
+
+Папка factory - папка с примером выполнения задания
+
+menu.ApplicationRunner - класс отвечающий за запуск приложения, после создание своей структуры необходимо
+добавить создание экземпляра своей организации в массив `organizations` так же как это уже сделанно в примере.
+
+Organization - абстрактный класс от которого нужно отнаследоваться для выполнения задания
+
+В этом задании необходимо:
+
+- Создать пакет с названием структуры. (Смотри пример в factory)
+- Создать базовую структуру и внести в нее все экземпляры взаимодействующий друг с другом объектов. (Смотри пример в factory)
+- Создать иерархию объектов которые будут взаимодействовать в структуре. (Смотри пример в factory)
+
+**Не забывайте проверять запускается ли программа перед отправкой на проверку!**
