--- src/objects/Mode.java
@@ -1,6 +1,0 @@
-package objects;
-
-public enum Mode {
-    OBJECT,
-    CALCULATOR
-}
