--- src/objects/Organization.java
@@ -1,5 +1,0 @@
-package objects;
-
-public abstract class Organization {
-    public abstract void start();
-}
