--- src/objects/calculator/Workplace.java
@@ -1,11 +1,0 @@
-package objects.calculator;
-
-public class Workplace {
-    public static void start() {
-        /*
-         * Здесь необходимо создать экземлпяр калькулятора и вывести в консоль
-         * все операции согласно readme.md в папке calculator
-         * */
-
-    }
-}
