--- src/objects/dns/Dns.java
@@ -1,34 +1,0 @@
-package objects.dns;
-
-import objects.Organization;
-import objects.dns.employees.Director;
-import objects.dns.employees.Employee;
-import objects.dns.employees.RegionalManager;
-
-public class Dns extends Organization {
-
-    @Override
-    public void start() {
-        Employee employee1 = new Employee("Иван", "Алексеев", 25, "Продавец-консультант");
-        Employee employee2 = new Employee("Роман", "Семенчук", 25, "Продавец-консультант");
-        Employee employee3 = new Employee("Дмитрий", "Акулов", 24, "Продавец-консультант");
-        Employee employee4 = new Employee("Валерий", "Крот", 27, "Продавец-консультант");
-        Employee employee5 = new Employee("Павел", "Попов", 33, "Старший продавец-консультант");
-        Employee employee6 = new Employee("Денис", "Шакин", 30, "Кладовщик");
-
-        Employee[] employees = {
-                employee1,
-                employee2,
-                employee3,
-                employee4,
-                employee5,
-                employee6
-        };
-
-        Director director = new Director("Борис", "Грицаев", 40, "Директор и царь всея отдела", employees);
-
-        RegionalManager regionalManager = new RegionalManager("СлаваБогу", "ТыПришел", 42, "Заведующий банхаммером ДНСа", director);
-
-        regionalManager.doWork();
-    }
-}
