--- src/objects/school/faculty/Human.java
@@ -1,21 +1,0 @@
-package objects.school.faculty;
-
-abstract class Human {
-    private String firstName;
-    private String lastName;
-    private int age;
-
-    Human(String firstName, String lastName, int age) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.age = age;
-    }
-
-    abstract void doWork();
-
-    void introduceYourself() {
-        System.out.println("I'm " + this.firstName + " " + this.lastName + "! I'm " + age + " years old");
-    }
-
-
-}
